chore: move func to utils to avoid cross dependency error
